DocumentQA/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pypdf==4.2.0
requests==2.32.3
werkzeug==3.0.3
groq